+++ src/main/scala/chisel3/testers/TesterDriver.scala
@@ -5,7 +5,30 @@ package chisel3.testers
 import chisel3._
 import java.io._
 
+import firrtl.{Driver => _, _}
+import _root_.util.BackendCompilationUtilities
+
 object TesterDriver extends BackendCompilationUtilities {
+
+  /**
+    * like 'firrtlToVerilog' except it runs the process inside the same JVM
+    *
+    * @param prefix basename of the file
+    * @param dir    directory where file lives
+    * @return       true if compiler completed successfully
+    */
+  def compileFirrtlToVerilog(prefix: String, dir: File): Boolean = {
+    val optionsManager = new ExecutionOptionsManager("chisel3") with HasChiselExecutionOptions with HasFirrtlOptions {
+      commonOptions = CommonOptions(topName = prefix, targetDirName = dir.getAbsolutePath)
+      firrtlOptions = FirrtlExecutionOptions(compilerName = "verilog")
+    }
+
+    firrtl.Driver.execute(optionsManager) match {
+      case _: FirrtlExecutionSuccess => true
+      case _: FirrtlExecutionFailure => false
+    }
+  }
+
   /** For use with modules that should successfully be elaborated by the
     * frontend, and which can be turned into executables with assertions. */
   def execute(t: () => BasicTester,
@@ -37,7 +60,7 @@ object TesterDriver extends BackendCompilationUtilities {
     if(runFirrtlasProcess) {
       // Use sys.Process to invoke a bunch of backend stuff, then run the resulting exe
       if ((firrtlToVerilog(target, path) #&&
-        verilogToCpp(target, target, path, additionalVFiles, cppHarness) #&&
+        verilogToCpp(target, path, additionalVFiles, cppHarness) #&&
           cppToExe(target, path)).! == 0) {
         executeExpectingSuccess(target, path)
       } else {
@@ -50,7 +73,7 @@ object TesterDriver extends BackendCompilationUtilities {
         return false
       }
       // Use sys.Process to invoke a bunch of backend stuff, then run the resulting exe
-      if ((verilogToCpp(target, target, path, additionalVFiles, cppHarness) #&&
+      if ((verilogToCpp(target, path, additionalVFiles, cppHarness) #&&
           cppToExe(target, path)).! == 0) {
         executeExpectingSuccess(target, path)
       } else {
